Refine last details, handle some edge cases
