+++ package.json
@@ -51,7 +51,7 @@
     "test-rpc": "cd scripts/tests && node test-rpc",
     "template": "cd customize.dist/src && for page in ../index.html ../contact.html ../features.html ../../www/login/index.html ../../www/register/index.html ../../www/user/index.html;do echo $page; cp template.html $page; done;",
     "evict-inactive": "node scripts/evict-inactive.js",
-    "build": "node build.js",
-    "clean-built": "rm -rf customize/www/"
+    "make-opengraph": "node scripts/build.js",
+    "clean-opengraph": "rm -rf customize/www/"
   }
 }
